Fix File.Move
